Use plugin mode for source generation in CMake
This commit uses protoc using the nanopb plugin to generate source files
in a single step instead of two. A new variable 'NANOPB_OPTIONS' can be
used to pass nanopb options to protoc.

The directory of a generated proto file is always passed as a search
path for option files.

This fixes the calculation of max encoded size for proto files with
imports.