Detect invalid wire type when decoding fields.
Previously nanopb only considered the type defined in .proto
when decoding fields. If the .proto in the encoding end didn't
match what was used in decoding end, e.g. varint field could get
decoded as e.g. prefix of a string field.

Mismatched field types in .proto is not expected to work correctly,
so in a sense this was garbage in, garbage out situation. However
adding these checks will make debugging easier when this happens.