fuzztest: workaround slowness with clang libsanitizer.
Currently nanopb dynamic allocation implementation grows the arrays with realloc()
as each new field arrives. This is the simplest and smallest code approach, but
results in a lot of realloc() calls. Normally this is a bit slow, but not hugely
so as typical realloc() implementation can grow the arrays in place.

However, libsanitizer realloc() implementation always copies the whole array.
This unnecessarily slows down fuzz testing.

This commit adds a rounding of the allocated block size in fuzztest to reduce
the number of realloc() calls. This does reduce the accuracy of libsanitizer tests
for large memory reallocations, but the custom checks in malloc_wrappers.c should
cover for the difference.