Add label_override field option
This lets proto3 files define submessage fields which nanopb treats as
required (even though proto3 doesn't support 'required'). This means that
nanopb won't generate a has_xxx field, will reject messages where this
field is missing, and will always serialize the field. However, other
consumers of the proto file won't see the field as required, and are
therefore safe to it being removed in future.

To do this, use:

    Submessage some_submessage = 1 [(nanopb).label_override = LABEL_REQUIRED];

This mechanism also lets users change whether nanopb treats the field as
optional or repeated in both proto2 and proto3.

Fixes: #962